backend/requirements.txt
numpy
Pillow
torch
torchvision
timm
segmentation-models-pytorch